refactor test dir
