+++ pytorch_blade/tests/disc/ops/test_batchnorm.py
@@ -12,7 +12,7 @@
 import torch
 import unittest
 
-from tests.mlir.testing_utils import DiscTestCase
+from tests.disc.testing_base import DiscTestCase
 
 
 class TestDiscBatchNorm(DiscTestCase):
