+++ pytorch_blade/tests/disc/ops/test_conv_ops.py
@@ -13,7 +13,7 @@ import torch
 import torch_blade
 import unittest
 
-from tests.mlir.testing_utils import DiscTestCase
+from tests.disc.testing_base import DiscTestCase
 
 class TestMlirConvolution(DiscTestCase):
     def _test_conv(self, conv_func, inp_test_data=None):
