+++ pytorch_blade/tests/disc/test_black_list.py
@@ -14,7 +14,7 @@ import unittest
 
 from torch.testing import FileCheck
 from torch_blade.config import Config
-from tests.mlir.testing_utils import DiscTestCase
+from tests.disc.testing_base import DiscTestCase
 
 
 class TestDiscBlackOps(DiscTestCase):
