+++ pytorch_blade/tests/disc/test_is_mlir_mhlo_supported.py
@@ -9,8 +9,27 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import torch
 import unittest
-from torch_blade.mlir import is_available
 
-def skipIfNoDISC():
-    return unittest.skipIf(not is_available(), "DISC support was not built")
+from torch_blade import mlir
+
+from tests.disc.testing_base import DiscTestCase
+
+
+class TestDiscTools(DiscTestCase):
+
+    def test_const_tensor(self):
+        gstr = """
+        graph(%x.1 : Tensor,
+              %y.1 : Tensor):
+          %4 : Tensor = aten::mul(%x.1, %y.1)
+          return (%4)"""
+
+        graph = torch._C.parse_ir(gstr)
+        all_unspt = all(not mlir.is_mlir_mhlo_supported(n) for n in graph.nodes())
+        self.assertTrue(all_unspt)
+
+
+if __name__ == "__main__":
+    unittest.main()
