+++ pytorch_blade/tests/disc/test_multi_streams.py
@@ -12,7 +12,7 @@
 import torch
 import unittest
 
-from tests.mlir.testing_utils import DiscTestCase
+from tests.disc.testing_base import DiscTestCase
 from torch_blade import version
 
 class Linear(torch.nn.Module):
